refactor: lite and rich text editors
