+++ packages/editor/lite-text-editor/src/extensions/index.tsx
@@ -1,3 +1,3 @@
-import { EnterKeyExtension } from "src/ui/extensions/enter-key-extension";
+import { EnterKeyExtension } from "src/extensions/enter-key-extension";
 
 export const LiteTextEditorExtensions = (onEnterKeyPress?: () => void) => [EnterKeyExtension(onEnterKeyPress)];
