+++ packages/editor/lite-text-editor/src/index.ts
@@ -1,7 +1,7 @@
-export { LiteTextEditor, LiteTextEditorWithRef } from "src/ui";
-export { LiteTextReadOnlyEditor, LiteTextReadOnlyEditorWithRef } from "src/ui/read-only";
+export { LiteTextEditor, LiteTextEditorWithRef } from "src/editor";
+export { LiteTextReadOnlyEditor, LiteTextReadOnlyEditorWithRef } from "src/read-only-editor";
 export type { IMentionSuggestion, IMentionHighlight } from "@plane/editor-core";
 
-export type { ILiteTextEditor } from "src/ui";
-export type { ILiteTextReadOnlyEditor } from "src/ui/read-only";
+export type { ILiteTextEditor } from "src/editor";
+export type { ILiteTextReadOnlyEditor } from "src/read-only-editor";
 export type { EditorRefApi, EditorReadOnlyRefApi, EditorMenuItem, EditorMenuItemNames } from "@plane/editor-core";
