+++ packages/editor/lite-text-editor/src/read-only-editor.tsx
