+++ packages/editor/rich-text-editor/src/index.ts
@@ -1,8 +1,8 @@
-export { RichTextEditor, RichTextEditorWithRef } from "src/ui";
-export { RichTextReadOnlyEditor, RichTextReadOnlyEditorWithRef } from "src/ui/read-only";
+export { RichTextEditor, RichTextEditorWithRef } from "src/editor";
+export { RichTextReadOnlyEditor, RichTextReadOnlyEditorWithRef } from "src/read-only-editor";
 
-export type { IRichTextEditor } from "src/ui";
+export type { IRichTextEditor } from "src/editor";
 
-export type { IRichTextReadOnlyEditor } from "src/ui/read-only";
+export type { IRichTextReadOnlyEditor } from "src/read-only-editor";
 export type { IMentionSuggestion, IMentionHighlight } from "@plane/editor-core";
 export type { EditorRefApi, EditorReadOnlyRefApi } from "@plane/editor-core";
