+++ packages/editor/rich-text-editor/src/read-only-editor.tsx
